feat(proto): widen ReplicaId to u16 across every wire surface
A replica index is a u16, decoupling replica-identity width from the u8
range. Every wire surface a replica id crosses widens with it:

- the message-body codec and each per-variant `encoded_len` replica
  contribution (U8 -> U16; the RequestSync recovery flag and the per-entry
  body-state tag stay u8);
- the two sync carrier overhead constants count the replica as 2 bytes, so
  `max_unchunked_snapshot_len` and `SYNC_CHUNK_LEN` stay byte-exact against
  the frame cap;
- the labeled hello encodes the replica id as 2 big-endian bytes (a replica
  hello is now 21 bytes; HELLO_VERSION bumped);
- the CertOid identity extension accepts a u16 replica id.

The three replica-keyed maps (peer_checkpoint, dvc_from, sync_serving) key
by ReplicaId so no call site truncates an id to u8. WIRE_VERSION bumped and
the affected codec/frame goldens re-pinned; codec, hello, and CertOid round
trips gain coverage for ids above the u8 range.

+++ viewstamp-proto/src/endpoint/recovery.rs
@@ -1244,7 +1244,7 @@ impl<S: StateMachine> Endpoint<S> {
     if self.pending_sb.is_some() {
       return;
     }
-    if m.replica().get() >= self.config.replica_count() {
+    if m.replica().get() >= self.config.replica_count() as u16 {
       return; // ignore malformed/out-of-range replica id
     }
     let (op, commit, floor, log) = if self.is_primary() {
